refactor(cost_estimator): split into sub-modules to respect 500 NLOC limit (#471)
cost_estimator.rs was 860 NLOC — 360 lines over the CONTRIBUTING.md
500 NLOC file limit (Devin Finding F on PR #606). Split the monolithic
file into a three-file module that keeps the public API intact via
re-exports and multi-file impl blocks:

- cost_estimator/mod.rs (306 NLOC)
  Public types (Cost, SelectivityMethod, CostEstimator),
  CostFactorsRef, COMPAT_* constants, and the base selectivity helpers
  (estimate_filter_cost, estimate_hnsw_search_cost_with_ef_on_size,
  estimate_condition_selectivity, comparison/in/between/like helpers,
  filter_cost_from_selectivity).

- cost_estimator/selectivity_method.rs (341 NLOC)
  Method-aware variants (estimate_condition_selectivity_with_method
  + per-predicate _with_method helpers) and the has_cardinality_data
  guard. Hosts the former selectivity_method_tests module (14 tests).

- cost_estimator/plan_cost.rs (236 NLOC)
  estimate_plan_cost + per-node cost functions (vector search, table
  scan, limit/offset, index lookup, match traversal). Hosts the
  former plan_cost_tests module (9 tests).

Public API unchanged: all external imports like
`use crate::velesql::cost_estimator::{CostEstimator, SelectivityMethod, Cost}`
continue to resolve because Rust folds directory-modules into the
same path. Multiple impl CostEstimator<'_> blocks across the three
files compile naturally (Rust allows inherent-impl fragmentation).

Tests: 23 tests pass (9 plan_cost + 14 selectivity_method).
